Resolve ties using the evaluation function
Negamax always assumes both players will never "make mistakes".  In at
least one case (XO-/-X-/OXO, X's turn) the algorithm wasn't choosing a
move that could result in a win if the opponent made a mistake, because
the overall score for perfect play in all cases was the same.  This adds
code to track when a number of moves have the same overall score, and
chooses the one that has the highest immediate score, so that if the
opponent makes a mistake, we can capitalize on it.