Add matches per tourney logic
Also, this changes how scoring works.  Now each time a score is tallied,
it's scaled down by how many total games there are in the tourney.  This
means that a net that wins x% of the time will score the same regardless
of how many matches it plays.

Also renames "Games" to "Matches".