Add unused age to individuals, simplify generation
I'm going to use the age parameter in the next iteration, but I wanted
to formalize its place in the code first.  This means we have to have a
function to order individuals (instead of using < or > on score), but
that's ok.

This also cleans up the generation code a little so it doesn't maintain
a members array that contains individuals, but simply the individuals
themselves.  I had originally intended the generation object to be the
only thing mucked around with by the workers (now, the individuals' own
score properties are set from the workers), but that's an arbitrary
distinction, and the end result is identical, so this aims for
simplicity.